+++ paper/src/main/java/cloud/grabsky/configuration/paper/adapter/SoundSourceAdapter.java
@@ -29,10 +29,10 @@ import net.kyori.adventure.sound.Sound;
 /**
  * Converts {@link String} to {@link Sound.Source} and vice-versa but using case-insensitive strategy.
  */
-public final class SoundSourceJsonAdapter extends AbstractEnumJsonAdapter<Sound.Source> {
-    /* SINGLETON */ public static final SoundSourceJsonAdapter INSTANCE = new SoundSourceJsonAdapter();
+public final class SoundSourceAdapter extends AbstractEnumJsonAdapter<Sound.Source> {
+    /* SINGLETON */ public static final SoundSourceAdapter INSTANCE = new SoundSourceAdapter();
 
-    private SoundSourceJsonAdapter() {
+    private SoundSourceAdapter() {
         super(Sound.Source.class, false);
     }
