+++ paper/src/main/java/cloud/grabsky/configuration/paper/adapter/StringComponentAdapter.java
@@ -24,15 +24,14 @@
 package cloud.grabsky.configuration.paper.adapter;
 
 import com.squareup.moshi.JsonAdapter;
+import com.squareup.moshi.JsonDataException;
 import com.squareup.moshi.JsonReader;
 import com.squareup.moshi.JsonReader.Token;
 import com.squareup.moshi.JsonWriter;
 import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
 import lombok.RequiredArgsConstructor;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.minimessage.MiniMessage;
-import org.checkerframework.checker.index.qual.PolyUpperBound;
 import org.jetbrains.annotations.NotNull;
 
 import java.io.IOException;
@@ -42,20 +41,15 @@ import java.util.List;
  * Converts {@link String} or {@link List List&lt;String&gt;} to {@link Component} using provided function.
  */
 @RequiredArgsConstructor(access = AccessLevel.PUBLIC)
-public final class ComponentJsonAdapter extends JsonAdapter<Component> {
-    /* DEFAULT */ public static final ComponentJsonAdapter INSTANCE = new ComponentJsonAdapter(MiniMessage.miniMessage());
+public final class StringComponentAdapter extends JsonAdapter<String> {
+    /* DEFAULT */ public static final StringComponentAdapter INSTANCE = new StringComponentAdapter(MiniMessage.miniMessage());
 
     private final MiniMessage miniMessage;
 
     @Override
-    public Component fromJson(final @NotNull JsonReader in) throws IOException {
+    public String fromJson(final @NotNull JsonReader in) throws IOException {
         if (in.peek() == Token.STRING) {
-            final String text = in.nextString();
-            // Returning empty component if value is null or blank
-            if ("".equals(text) == true)
-                return Component.empty();
-            // Parsing and returning
-            return miniMessage.deserialize(text);
+            return in.nextString();
         } else if (in.peek() == Token.BEGIN_ARRAY) {
             final StringBuilder builder = new StringBuilder();
             // ...
@@ -73,15 +67,15 @@ public final class ComponentJsonAdapter extends JsonAdapter<Component> {
             // ...
             in.endArray();
             // ...
-            return miniMessage.deserialize(builder.toString());
+            return builder.toString();
         } else {
             in.skipValue();
         }
-        return Component.empty();
+        throw new JsonDataException("Expected STRING or BEGIN_ARRAY, but found " + in.peek() + " at $." + in.getPath());
     }
 
     @Override
-    public void toJson(final @NotNull JsonWriter out, final Component value) {
+    public void toJson(final @NotNull JsonWriter out, final String value) {
         throw new UnsupportedOperationException("NOT IMPLEMENTED");
     }
 }
